# pay: Speed up `pay` and fix the shadow route construction