@rustyrussell 2020-04-02
As kids say these days, "LOL WUT?"

Ack 79c05b0c48f3f210b3fc999c79d6cd219f6eacaa