@rustyrussell 2020-04-02
Also deserves a Changelog-Fixed: line IMHO.