@cdecker 2020-04-03
Added two changelog entries: one for the speedup and one for the shadow route fix :wink: